% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gadget_gatingTree.R
\name{gatingTreeApp}
\alias{gatingTreeApp}
\title{gatingTree shiny app}
\usage{
gatingTreeApp(gs)
}
\arguments{
\item{gs}{A GatingSet}
}
\value{
A shiny app object
}
\description{
This app allows you to interactively plot gated data by clicking the tree note
}
\examples{
if (interactive()) {
  app <- gatingTreeApp(gs)
  runApp(app)
}
}

