% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gadget_gatingTree.R
\name{gatingTreeGadget}
\alias{gatingTreeGadget}
\title{gatingTree gadget}
\usage{
gatingTreeGadget(gs)
}
\arguments{
\item{gs}{A GatingSet}
}
\description{
This gadget allows you to interactively plot gated data
}
\examples{
if (interactive()) {
  gatingTreeGadget(gs)
}
}

