% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/module_gatingMethod.R
\name{genericArgsUI}
\alias{genericArgsUI}
\title{Generic UI module for the gating method that does not have dedicated controls defined.}
\usage{
genericArgsUI(id)
}
\description{
Generic UI module for the gating method that does not have dedicated controls defined.
}

